requirements.txt
requests~=2.27.1
python-dotenv~=0.19.2
playwright~=1.18.2
discord~=1.7.3